Implement first iteration of the controller loop
The user is expected to provide a ConfigMap containing the
homeserver.yaml configuration file for Synapse. The controller runs
basic validation checks against the configuration file and retrieve the
"server_name" and "report_stats" value, needed at a later step for the
generation of the missing config files.

The controller first creates a PVC for hosting the application's data.

The controller then creates a deployment for Synapse. This deployment
generates missing config files (key file and log config file) using the
"generate" argument of the synapse container in an initContainer.

Finally, the controller creates a Service to allows connection to the
Synapse server.

Limitations:
- The current implementation assumes that the path of the key file and
  of the log config file are in /data
- In Openshift the default SA must have the privileged SCC.