Enable option to provision a PostgreSQL DB
This commits enables the spec.CreateNewPostgreSQL attribute (previsouly
present but unused). When set to true, the Synapse Operator will
automatically:
* Provision a PostgreSQL Cluster
* Create a "synapse" database correctly configured for the "synapse"
user.
* Register the database connection information into
spec.status.databaseConnectionInfo
* Edit the homeserver.yaml configuration file - inside the ConfigMap given by
spec.HomeserverConfigMapName - to configure the synapse instance to
connect to the newly created Database.

This creates a dependency from the Synapse Operator to the Crunchy
postgres-operator. This dependency is currenly not checked.

Closes #9