Allow users to provide configuration values
So far, users needed to create a ConfigMap from a homeserver.yaml
configuration file. Starting with this commit, this is not a hard
pre-requisite anymore. Users can instead pass the configuration values
via the Spec.Config.Values section of the Synapse CRD. The Synapse
controller takes care of crafting a configuation file based on those
values and a ConfigMap to store it.

Closes #13