Enable registration by default in Synapse
In most cases, users who deploy a Synapse instance expect to be able to
be create a account. This also makes it easier for testing purposes.

We could also add this configuration option under Spec.Config.Values if
there is the need for setting up Synapse instances with registration
disabled.