Add option to deploy the mautrix-signal bridge
It is composed of:
* mautrix-signal: Deployement, Service, 1 PVC to host the bridge's data,
1 PVC to host the SQLite database file, and a ConfigMap to provide the
input config.yaml
* signald: Deployment and PVC

As for the configuration of Synapse (homeserver.yaml), and Heisenbridge
(heisenbridge.yaml), the user can provide its own config.yaml for
mautrix-signal. In that case, a copy of the user-provided ConfigMap is
created. In the copied object, some configuration options are
overwritten. The user-provided ConfigMap is not modified.

Alternatively, if the user doesn't provide a ConfigMap, a default
config.yaml is used for configuring the mautrix-signal bridge.

The controller also takes care of updating the configuration of Synapse
so that it is aware of the newly deployed bridge (added as an appservice
in homeserver.yaml)