Use template files and embed.FS instead of go structs
Instead of describing the basis for the child resources
as a Go struct, which requires a LOT of indentation, makeing it less
readable, uses an actual yaml template, and embeds that file into the
binary. Then, it uses an embed.FS to read that file from the "filesystem"
and parse it into an actual go struct. Then the rest
of the code just operates as it did before.

Signed-off-by: Brad P. Crochet <brad@redhat.com>